chore: remove zen_go_bindings
