+++ Makefile
@@ -3,7 +3,7 @@ prepare: check_binaries
 	git submodule update --remote --merge
 	mkdir -p /opt/aikido/lib
 	cp .cache/binaries/* /opt/aikido/lib/
-	cd agent/ && make install_protoc go_setup build
+	cd agent/ && make install_protoc go_setup
 
 .PHONY: test
 test: prepare
